cleanup(core): move migrate run-config resolvers to shared modules
The run/ worker must not import migrate.ts (the module boundary forbids
the cycle), so the helpers that resolve a run's effective configuration
move next to their subjects: resolveCreateCommits and
confirmCommitsOnDefaultBranch into migrate-commits.ts,
resolveShouldRunValidation into agentic/select.ts, and
resolveAgenticRunId into agentic/handoff.ts. Verbatim moves; migrate.ts
consumes them from their new homes.
