+++ packages/nx/src/command-line/migrate/agentic/select.ts
@@ -30,8 +30,9 @@ export interface ResolveAgenticInput {
 }
 
 /**
- * Resolves the agentic state for a `--run-migrations` invocation. Runs once,
- * before the migration loop, and its result is cached for every entry.
+ * Resolves the agentic state for a run-phase invocation (`--run-migrations`
+ * or `--run-migration`). Runs once, before any migration executes, and its
+ * result is cached for every entry.
  */
 export async function resolveAgentic(
   input: ResolveAgenticInput
@@ -82,6 +83,22 @@ export async function resolveAgentic(
   }
 
   return { kind: 'enabled', selectedAgent: selected };
+}
+
+/**
+ * Resolves whether the framework-owned generic-validation agent step should run
+ * after generator-only migrations.
+ *
+ * Default-on when the agentic flow resolved to `enabled`; silently ignored
+ * otherwise (no warning emitted): `--validate` requires an active agent
+ * session by definition. An explicit `--no-validate` (`validate === false`)
+ * opts out even when agentic is enabled.
+ */
+export function resolveShouldRunValidation(args: {
+  validate: boolean | undefined;
+  agenticKind: ResolvedAgentic['kind'];
+}): boolean {
+  return args.validate !== false && args.agenticKind === 'enabled';
 }
 
 /**
