+++ packages/nx/src/command-line/migrate/migrate-commits.ts
@@ -2,6 +2,8 @@ import * as pc from 'picocolors';
 import { hasUncommittedChanges, tryCommitChanges } from '../../utils/git-utils';
 import { logger } from '../../utils/logger';
 import { output } from '../../utils/output';
+import type { ResolvedAgentic } from './agentic/types';
+import { migratePrompt } from './safe-prompt';
 
 /**
  * Discriminated result for `commitMigrationIfRequested`. Distinguishes the
@@ -147,4 +149,109 @@ export function commitCheckpointBeforeMigrations(
       ],
     });
   }
+}
+
+/**
+ * Resolves the effective `--create-commits` state once the agentic flow has
+ * been resolved. When per-migration commits do not isolate a migration's
+ * diff, the agent's outer prompt falls back to an embedded file list instead
+ * of pointing at git; the diff-context flag returned here gates that choice.
+ */
+export function resolveCreateCommits(args: {
+  createCommits: boolean | undefined;
+  agenticKind: ResolvedAgentic['kind'];
+  isGitRepo: boolean;
+  /**
+   * Whether `--commit-prefix` was given a non-default value. When commits
+   * end up disabled, the prefix has no effect. The warning copy below
+   * surfaces that so the user isn't silently misled.
+   */
+  commitPrefixIsCustom?: boolean;
+}): {
+  effective: boolean;
+  agenticHasDiffContext: boolean;
+  warning?: string;
+  error?: string;
+} {
+  const { createCommits, agenticKind, isGitRepo, commitPrefixIsCustom } = args;
+
+  // Explicit `--create-commits` without git is a hard error: the user asked
+  // for something we cannot deliver.
+  if (createCommits === true && !isGitRepo) {
+    return {
+      effective: false,
+      agenticHasDiffContext: false,
+      error:
+        '`--create-commits` requires a git repository. Run `git init` first, or omit the flag.',
+    };
+  }
+
+  if (agenticKind === 'enabled') {
+    if (createCommits === false) {
+      return {
+        effective: false,
+        agenticHasDiffContext: false,
+        warning:
+          "--no-create-commits was passed alongside --agentic. Without per-migration commits, the agent can't isolate the current migration's changes from earlier migrations in this run. Drop --no-create-commits for accurate per-migration review." +
+          (commitPrefixIsCustom
+            ? ' Note: the custom --commit-prefix value will have no effect because commits are disabled.'
+            : ''),
+      };
+    }
+    // Without git we cannot soft-force commits the user didn't explicitly
+    // opt into. Degrade rather than error: continue the agentic run, but
+    // without per-file diff context (which depends on per-migration commits).
+    if (!isGitRepo) {
+      return {
+        effective: false,
+        agenticHasDiffContext: false,
+        warning:
+          '`--agentic` enables per-migration commits by default, but the workspace is not a git repository. Continuing without commits, so the agent will not receive per-file diff context. Run `git init` to enable.' +
+          (commitPrefixIsCustom
+            ? ' The custom --commit-prefix value will have no effect.'
+            : ''),
+      };
+    }
+    return { effective: true, agenticHasDiffContext: true };
+  }
+
+  // Commits aren't enabled here. A custom prefix (from --commit-prefix or
+  // nx.json) would have no effect; surface that instead of dropping it
+  // silently.
+  return {
+    effective: createCommits === true,
+    agenticHasDiffContext: false,
+    warning:
+      commitPrefixIsCustom && createCommits !== true
+        ? 'A custom migrate commit prefix is configured, but commits are not enabled for this run, so it has no effect. Set `migrate.createCommits` to `true` (or pass `--create-commits`) to create a commit per migration.'
+        : undefined,
+  };
+}
+
+/**
+ * Confirms before creating per-migration commits while the user sits on the
+ * repository's default branch, so migrations don't silently commit there (most
+ * relevant since `--agentic` enables commits by default). Only the default
+ * branch triggers a prompt; a detached HEAD, a different branch, or an
+ * unresolved default branch all proceed untouched. Callers gate this on
+ * commits being effective and prompting being possible, so non-interactive
+ * runs (CI, `--no-interactive`) never reach here.
+ */
+export async function confirmCommitsOnDefaultBranch(args: {
+  currentBranch: string | null;
+  defaultBranch: string | null;
+}): Promise<boolean> {
+  const { currentBranch, defaultBranch } = args;
+  if (!currentBranch || !defaultBranch || currentBranch !== defaultBranch) {
+    return true;
+  }
+  const { proceed } = await migratePrompt<{ proceed: boolean }>([
+    {
+      name: 'proceed',
+      type: 'confirm',
+      message: `You're on the default branch '${currentBranch}'. nx migrate will create a commit for each migration on this branch. Continue?`,
+      initial: false,
+    },
+  ]);
+  return proceed;
 }
