+++ packages/nx/src/command-line/migrate/migrate.ts
@@ -12,7 +12,6 @@ import {
   lt,
   lte,
   major,
-  rsort,
   satisfies,
   valid,
 } from 'semver';
@@ -139,6 +138,8 @@ import { applyAgenticHandoffGitignoreFallback } from './agentic/handoff-gitignor
 import {
   commitCheckpointBeforeMigrations,
   commitMigrationIfRequested,
+  confirmCommitsOnDefaultBranch,
+  resolveCreateCommits,
 } from './migrate-commits';
 import {
   buildDirectiveBlockBodyLines,
@@ -2585,10 +2586,6 @@ function showConnectToCloudMessage() {
 
 export { isPromptOnlyMigration, isHybridMigration };
 
-export function resolveAgenticRunId(migrations: ExecutableMigration[]): string {
-  return rsort(migrations.map((m) => normalizeVersion(m.version)))[0]!;
-}
-
 export function formatSkippedPromptsNextStep(
   skipped: PlannedMigration[]
 ): string {
@@ -2596,128 +2593,6 @@ export function formatSkippedPromptsNextStep(
     'Some prompt migrations were skipped. Review and apply each of the following prompt files to the workspace, in the listed order:',
     ...skipped.map((m) => `  - ${m.prompt}`),
   ].join('\n');
-}
-
-/**
- * Resolves the effective `--create-commits` state once the agentic flow has
- * been resolved. The agent's outer prompt only embeds the impl-phase file list
- * when per-migration commits isolate each migration's diff, so the diff-context
- * flag returned here gates that section.
- */
-export function resolveCreateCommits(args: {
-  createCommits: boolean | undefined;
-  agenticKind: ResolvedAgentic['kind'];
-  isGitRepo: boolean;
-  /**
-   * Whether `--commit-prefix` was given a non-default value. When commits
-   * end up disabled, the prefix has no effect — the warning copy below
-   * surfaces that so the user isn't silently misled.
-   */
-  commitPrefixIsCustom?: boolean;
-}): {
-  effective: boolean;
-  agenticHasDiffContext: boolean;
-  warning?: string;
-  error?: string;
-} {
-  const { createCommits, agenticKind, isGitRepo, commitPrefixIsCustom } = args;
-
-  // Explicit `--create-commits` without git is a hard error — the user asked
-  // for something we cannot deliver.
-  if (createCommits === true && !isGitRepo) {
-    return {
-      effective: false,
-      agenticHasDiffContext: false,
-      error:
-        '`--create-commits` requires a git repository. Run `git init` first, or omit the flag.',
-    };
-  }
-
-  if (agenticKind === 'enabled') {
-    if (createCommits === false) {
-      return {
-        effective: false,
-        agenticHasDiffContext: false,
-        warning:
-          "--no-create-commits was passed alongside --agentic. Without per-migration commits, the agent can't isolate the current migration's changes from earlier migrations in this run. Drop --no-create-commits for accurate per-migration review." +
-          (commitPrefixIsCustom
-            ? ' Note: the custom --commit-prefix value will have no effect because commits are disabled.'
-            : ''),
-      };
-    }
-    // Without git we cannot soft-force commits the user didn't explicitly
-    // opt into. Degrade rather than error: continue the agentic run, but
-    // without per-file diff context (which depends on per-migration commits).
-    if (!isGitRepo) {
-      return {
-        effective: false,
-        agenticHasDiffContext: false,
-        warning:
-          '`--agentic` enables per-migration commits by default, but the workspace is not a git repository. Continuing without commits — the agent will not receive per-file diff context. Run `git init` to enable.' +
-          (commitPrefixIsCustom
-            ? ' The custom --commit-prefix value will have no effect.'
-            : ''),
-      };
-    }
-    return { effective: true, agenticHasDiffContext: true };
-  }
-
-  // Commits aren't enabled here. A custom prefix only reaches this path via
-  // nx.json (e.g. `migrate.commitPrefix` + `migrate.agentic` when the agentic
-  // flow resolves to disabled); surface that it has no effect rather than
-  // dropping it silently.
-  return {
-    effective: createCommits === true,
-    agenticHasDiffContext: false,
-    warning:
-      commitPrefixIsCustom && createCommits !== true
-        ? 'A custom migrate commit prefix is configured, but commits are not enabled for this run, so it has no effect. Set `migrate.createCommits` to `true` (or pass `--create-commits`) to create a commit per migration.'
-        : undefined,
-  };
-}
-
-/**
- * Confirms before creating per-migration commits while the user sits on the
- * repository's default branch, so migrations don't silently commit there (most
- * relevant since `--agentic` enables commits by default). Only the default
- * branch triggers a prompt; a detached HEAD, a different branch, or an
- * unresolved default branch all proceed untouched. Callers gate this on
- * commits being effective and prompting being possible, so non-interactive
- * runs (CI, `--no-interactive`) never reach here.
- */
-export async function confirmCommitsOnDefaultBranch(args: {
-  currentBranch: string | null;
-  defaultBranch: string | null;
-}): Promise<boolean> {
-  const { currentBranch, defaultBranch } = args;
-  if (!currentBranch || !defaultBranch || currentBranch !== defaultBranch) {
-    return true;
-  }
-  const { proceed } = await migratePrompt<{ proceed: boolean }>([
-    {
-      name: 'proceed',
-      type: 'confirm',
-      message: `You're on the default branch '${currentBranch}'. nx migrate will create a commit for each migration on this branch. Continue?`,
-      initial: false,
-    },
-  ]);
-  return proceed;
-}
-
-/**
- * Resolves whether the framework-owned generic-validation agent step should run
- * after generator-only migrations.
- *
- * Default-on when the agentic flow resolved to `enabled`; silently ignored
- * otherwise (no warning emitted) — `--validate` requires an active agent
- * session by definition. An explicit `--no-validate` (`validate === false`)
- * opts out even when agentic is enabled.
- */
-export function resolveShouldRunValidation(args: {
-  validate: boolean | undefined;
-  agenticKind: ResolvedAgentic['kind'];
-}): boolean {
-  return args.validate !== false && args.agenticKind === 'enabled';
 }
 
 export async function executeMigrations(
@@ -2747,7 +2622,7 @@ export async function executeMigrations(
       }
     | undefined;
   if (agentic?.kind === 'enabled' && sortedMigrations.length > 0) {
-    const { initRunDir } =
+    const { initRunDir, resolveAgenticRunId } =
       require('./agentic/handoff') as typeof import('./agentic/handoff');
     const { runAgenticPromptStep } =
       require('./agentic/run-step') as typeof import('./agentic/run-step');
@@ -3223,7 +3098,7 @@ async function runMigrations(
     migrationCount: migrations.length,
   });
 
-  const { resolveAgentic } =
+  const { resolveAgentic, resolveShouldRunValidation } =
     require('./agentic/select') as typeof import('./agentic/select');
   let agentic: ResolvedAgentic;
   try {
